added user roles

+++ app/controllers/wikis_controller.rb
@@ -1,7 +1,8 @@
 class WikisController < ApplicationController
-  
+
   def index
     @wikis = Wiki.all
+    authorize @wikis
   end
 
   def show
@@ -10,52 +11,53 @@ class WikisController < ApplicationController
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
-  
+
   def create
+    @wiki = Wiki.new(params.require(:wiki).permit(:title, :body, :private))
+    authorize @wiki
 
-     @wiki = Wiki.new
-     @wiki.title = params[:wiki][:title]
-     @wiki.body = params[:wiki][:body]
-
-     if @wiki.save
-
-       flash[:notice] = "wiki was saved."
-       redirect_to @wiki
-     else
-       flash.now[:alert] = "There was an error saving the wiki. Please try again."
-       render :new
-     end
+    if @wiki.save
+      flash[:notice] = "Wiki has been saved."
+      redirect_to @wiki
+    else
+      flash[:notice] = "There was an error saving the wiki. Please try again."
+      render :new
+    end
   end
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
-  
-    def update
-     @wiki = Wiki.find(params[:id])
-     @wiki.title = params[:wiki][:title]
-     @wiki.body = params[:wiki][:body]
- 
-     if @wiki.save
-       flash[:notice] = "Post was updated."
-       redirect_to @wiki
-     else
-       flash.now[:alert] = "There was an error saving the post. Please try again."
-       render :edit
-     end
+
+  def update
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
+
+    if @wiki.update_attributes(params.require(:wiki).permit(:title, :body, :private))
+      flash[:notice] = "Wiki has been updated."
+      redirect_to @wiki
+    else
+      flash[:error] = "There was an error updating the wiki. Please try again."
+      render :edit
     end
-    
-    def destroy
-     @wiki = Wiki.find(params[:id])
- 
-     if @wiki.destroy
-       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-       redirect_to wikis_path
-     else
-       flash.now[:alert] = "There was an error deleting the post."
-       render :show
-     end
+  end
+
+  def destroy
+    @wiki = Wiki.find(params[:id])
+    title = @wiki.title
+    authorize @wiki
+
+    if @wiki.destroy
+      flash[:notice] = "\"#{title}\" was deleted successfully."
+      redirect_to @wiki
+    else
+      flash[:error] = "There was an error deleting the wiki."
+      render :show
     end
-  
+
+  end
+
 end
